requirements.txt
langchain[all]
openai
streamlit
python-dotenv
pydantic==1.10.9
gpt4all